+++ migrations/2017_05_10_015415_create_h5p_counters_table.php
@@ -17,7 +17,7 @@ class CreateH5pCountersTable extends Migration
             $table->string('type', 63);
             $table->string('library_name', 127);
             $table->string('library_version', 31);
-            $table->integer('num')->unsigned();
+            $table->bigInteger('num')->unsigned();
             $table->primary(['type', 'library_name', 'library_version'], 'fk_primary');
         });
     }
