+++ migrations/2017_05_10_015415_create_h5p_libraries_languages_table.php
@@ -14,7 +14,7 @@ class CreateH5pLibrariesLanguagesTable extends Migration
     public function up()
     {
         Schema::create('h5p_libraries_languages', function (Blueprint $table) {
-            $table->integer('library_id')->unsigned();
+            $table->bigInteger('library_id')->unsigned();
             $table->string('language_code', 31);
             $table->text('translation', 65535);
             $table->primary(['library_id', 'language_code'], 'fk_primary');
